+++ src/canister_types/src/models/transaction_data.rs
@@ -1,7 +1,9 @@
 use candid::{CandidType, Nat, Principal};
-use icrc_ledger_types::icrc1::transfer::Memo;
+use icrc_ledger_types::{icrc::generic_metadata_value::MetadataValue, icrc1::transfer::Memo};
 use serde::{Deserialize, Serialize};
 
+// These structs is used on the transaction_handler cansister
+// changes to these structs should be reflected on the transaction_handler canister
 #[derive(CandidType, Deserialize, Serialize, Clone, Debug)]
 pub struct TransactionData {
     pub sender: Principal,
@@ -11,4 +13,14 @@ pub struct TransactionData {
     pub fee: Nat,
     pub block_height: Nat,
     pub amount: Nat,
+    pub metadata: Vec<(String, MetadataValue)>,
+}
+
+#[derive(CandidType, Deserialize, Serialize, Clone, Debug)]
+pub struct TransactionCompleteData {
+    pub sender: Principal,
+    pub receiver_count: u64,
+    pub total_amount_distributed: Nat,
+    pub canister: Principal,
+    pub metadata: Vec<(String, MetadataValue)>,
 }